requirements.txt
agno>=1.0.0
chromadb>=0.4.0
openai>=1.0.0
anthropic>=0.30.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
click>=8.1.0
rich>=13.7.0
loguru>=0.7.0
